Created backend_store under chain and moved db_contextXXX files into it.
